Modificaciones finales para la version 1.0
El código ya funciona, y a este punto quería llegar... Ahora, debes darte cuenta de por qué no podemos usar este enfoque y qué pequeña modificación podemos hacer para que
no se aprecien las transiciones.

Pista: Hay dos opciones sencillas, una de ellas involucra tener en cuenta dos códigos simultáneamente, otra un periodo de guarda.

+++ writeFrameToFinalVideo.m
@@ -1,3 +1,3 @@
 function writeFrameToFinalVideo(video, frame)
-    writeVideo(video, frame);
+    writeVideo(video, uint8(frame));
 end
